Rename "LiveObjects" file names to just "Objects"
