+++ objects.d.ts
