+++ src/plugins/objects/statemessage.ts
