+++ test/common/modules/objects_helper.js
