+++ test/package/browser/template/server/resources/index-objects.html
@@ -5,7 +5,7 @@
     <title>Ably NPM package test (Objects plugin export)</title>
   </head>
   <body>
-    <script type="text/javascript" src="index-liveobjects.js"></script>
+    <script type="text/javascript" src="index-objects.js"></script>
     <script type="text/javascript" src="runTest.js"></script>
   </body>
 </html>
