+++ test/package/browser/template/src/index-objects.ts
